Refactor/increase encapsulation of comm link handling (#40)
Refactor/increase encapsulation of comm link handling 
  * Encasulate MessageTimes dict into MessageTimes object in link.py and refactor timeout processing to be more simpler and more direct
  * refactor timeout processing to be more simpler and more direct
  * Move ack tracking to AckManager
  * Allow darglint to pass
  * Refactor: put link related code in its own package
  * Re-add link_state.py with one import to preserve backwards compatibility
